+++ qtensor/qtensor_op.cpp
@@ -1263,7 +1263,7 @@ void svd(qtensor<T>& A,
     int64_t ns;
     T* s_data;
     _S.read_all(&ns,&s_data);
-    for (size_t j = 0; j < ns; j++) {
+    for (int64_t j = 0; j < ns; j++) {
       S_and_block_idx.push_back(std::make_pair(s_data[j], ii));
       nm += s_data[j]*s_data[j];
     }
@@ -1428,13 +1428,13 @@ void svd(qtensor<T>& A,
 
   }//end midQ loop*/
 #ifndef NDEBUG
-  for(int ii=0;ii<mid_Q.size();ii++){
+  for(size_t ii=0;ii<mid_Q.size();ii++){
     assert(U.rank==U._block[ii].order);
-    for(int l=0;l<U.rank;l++){
+    for(unsigned l=0;l<U.rank;l++){
       assert(U._block[ii].lens[l]==U.block_index_qd[ii][l]);
     }
     assert(V.rank==V._block[ii].order);
-    for(int l=0;l<V.rank;l++){
+    for(unsigned l=0;l<V.rank;l++){
       assert(V._block[ii].lens[l]==V.block_index_qd[ii][l]);
     }
   }
